util: remove Parallel for TestIndexRange
